[Fix] Do not display internal branches (suffixed by -localehub)
